requirements.txt
Django
Django-grappelli
Django-filebrowser
Pillow
django-photologue
pytz
dj-static